refactor: Remove MeasurementGuide component
